Moved the enums into a custom module
